--- activity/desktop.ini
@@ -1,6 +1,0 @@
-[LocalizedFileNames]
-jd_wxFans.js=@jd_wxFans.js,0
-jd_super_mh.js=@jd_super_mh.js,0
-jd_mx_shop.js=@jd_mx_shop.js,0
-jd_mohe.js=@jd_mohe.js,0
-jd_priceProtect.js=@jd_priceProtect.js,0
