+++ jd_speed_redpocke.js
@@ -80,7 +80,7 @@ if ($.isNode()) {
 async function jsRedPacket() {
   try {
     await invite();
-    await sign();//极速版签到提现
+    //await sign();//极速版签到提现
     await reward_query();
     for (let i = 0; i < 3; ++i) {
       await redPacket();//开红包
